chore: update entrypoint for esp-idf

+++ esp-idf/entrypoint.sh
@@ -10,7 +10,6 @@ export ESPTOOLPY_NO_STUB=1
 FLASH_PORT="rfc2217://host.docker.internal:${ESP_RFC2217_PORT}?ign_set_control"
 
 sleep 2
-python3 /usr/local/bin/generate_kconfig.py 
 cd /workspace
 if [ "$1" = "lint" ]; then
     echo "Running cppcheck..."
@@ -18,9 +17,12 @@ if [ "$1" = "lint" ]; then
 
 # if menuconfig
 elif [ "$1" = "menuconfig" ]; then
-    # IF SDKCONFIG DOES EXIST REMOVE IT
-    if [ -f sdkconfig ]; then
-        rm sdkconfig
+    python3 /usr/local/bin/generate_kconfig.py 
+    if [ "$2" = "clean" ]; then
+        echo "Cleaning build directory..."
+        if [ -f sdkconfig ]; then
+            rm sdkconfig
+        fi
     fi
     idf.py menuconfig
 elif [ $# -eq 0 ]; then
